Rename     ManaPHP/Cache/Adapter/Apcu.php (from ManaPHP/Cache/Engine/Apcu.php) Rename     ManaPHP/Cache/Adapter/Db.php (from ManaPHP/Cache/Engine/Db.php) Rename     ManaPHP/Cache/Adapter/Db/Model.php (from ManaPHP/Cache/Engine/Db/Model.php) Rename     ManaPHP/Cache/Adapter/Db/manaphp_cache.sql (from ManaPHP/Cache/Engine/Db/manaphp_cache.sql) Rename     ManaPHP/Cache/Adapter/File.php (from ManaPHP/Cache/Engine/File.php) Rename     ManaPHP/Cache/Adapter/Memory.php (from ManaPHP/Cache/Engine/Memory.php) Rename     ManaPHP/Cache/Adapter/Redis.php (from ManaPHP/Cache/Engine/Redis.php) Deleted    ManaPHP/Cache/EngineInterface.php Rename     tests/CacheAdapterApcuTest.php (from tests/CacheEngineApcuTest.php) Rename     tests/CacheAdapterDbTest.php (from tests/CacheEngineDbTest.php) Rename     tests/CacheAdapterFileTest.php (from tests/CacheEngineFileTest.php) Rename     tests/CacheAdapterMemoryTest.php (from tests/CacheEngineMemoryTest.php) Rename     tests/CacheAdapterRedisTest.php (from tests/CacheEngineRedisTest.php) Modified   ManaPHP/Cache.php Modified   ManaPHP/Mvc/Factory.php Modified   tests/CacheTest.php

+++ tests/CacheAdapterApcuTest.php
@@ -1,33 +1,23 @@
 <?php
+
 namespace Tests;
 
-use ManaPHP\Cache\Engine\Db;
-use ManaPHP\Db\Adapter\Mysql;
-use ManaPHP\Di\FactoryDefault;
+use ManaPHP\Cache\Adapter\Apcu;
 use PHPUnit\Framework\TestCase;
 
-class CacheEngineDbTest extends TestCase
+class CacheAdapterApcuTest extends TestCase
 {
-    public function setUp()
-    {
-        $di = new FactoryDefault();
-
-        $di->setShared('db', function () {
-            $config = require __DIR__ . '/config.database.php';
-            $db = new Mysql($config['mysql']);
-            $db->attachEvent('db:beforeQuery', function (\ManaPHP\DbInterface $source, $data) {
-                //  var_dump(['sql'=>$source->getSQL(),'bind'=>$source->getBind()]);
-                var_dump($source->getSQL(), $source->getEmulatedSQL(2));
-
-            });
-            return $db;
-        });
-    }
-
+    /**
+     * @requires  extension apc
+     */
     public function test_exists()
     {
-        $cache = new Db();
+        if (!function_exists('apcu_exists')) {
+            $this->markTestSkipped();
+            return;
+        }
 
+        $cache = new Apcu();
         $cache->delete('var');
         $this->assertFalse($cache->exists('var'));
         $cache->set('var', 'value', 1000);
@@ -36,8 +26,12 @@ class CacheEngineDbTest extends TestCase
 
     public function test_get()
     {
-        $cache = new Db();
+        if (!function_exists('apcu_exists')) {
+            $this->markTestSkipped();
+            return;
+        }
 
+        $cache = new Apcu();
         $cache->delete('var');
 
         $this->assertFalse($cache->get('var'));
@@ -47,7 +41,12 @@ class CacheEngineDbTest extends TestCase
 
     public function test_set()
     {
-        $cache = new Db();
+        if (!function_exists('apcu_exists')) {
+            $this->markTestSkipped();
+            return;
+        }
+
+        $cache = new Apcu();
 
         $cache->set('var', '', 100);
         $this->assertSame('', $cache->get('var'));
@@ -62,12 +61,20 @@ class CacheEngineDbTest extends TestCase
         $cache->set('var', 'value', 1);
         $this->assertTrue($cache->exists('var'));
         sleep(2);
-        $this->assertFalse($cache->exists('var'));
+        /**
+         * After the ttl has passed, the stored variable will be expunged from the cache (on the next request). If no ttl is supplied (or if the ttl is 0), the value will persist until it is removed from the cache manually, or otherwise fails to exist in the cache (clear, restart, etc.).
+         */
+        $this->assertTrue($cache->exists('var'));
     }
 
     public function test_delete()
     {
-        $cache = new Db();
+        if (!function_exists('apcu_exists')) {
+            $this->markTestSkipped();
+            return;
+        }
+
+        $cache = new Apcu();
 
         //exists and delete
         $cache->set('var', 'value', 100);
